fix(train): split regex.rs and its tests for the file cap; classify the new roots
Two files went over 2000 lines:

- regex.rs -> regex/compile_cache.rs takes the program-compilation and
  cache group (size limit, std/fancy builders, eviction, the checked
  compile-and-cache entry). `js_regexp_new_impl` was the obvious bigger
  extraction and is deliberately NOT the one taken: it carries the two
  #7341 raw-handle debt sites, and raw_handle_debt.py's --no-raise-vs arm
  refuses a ceiling on a file absent at the merge base, so moving it would
  have forced either surgery on the allocation path two perf PRs are
  tuning, or a with_const_ptr(|p| p) that games the ratchet without
  scoping anything.
- regex/tests.rs -> regex/tests_part2.rs, split at a test boundary. All
  70 #[test] items are accounted for across the two files, and part2
  carries the same cfg(all(test, feature)) gate as its sibling.

#9893's new roots are classified: REGEXP_PROTOTYPE_PTR_SLOT and
REGEXP_PROTOTYPE_TEST_CLOSURE_SLOT are covered_elsewhere, naming
scan_object_cache_roots_mut, which really does visit them (as an i64 slot
and a nanbox word respectively) and is registered via reg_scanner!. The
test index and the walk counter are not_a_gc_pointer.

NEVER_MATCH moved with the compile-cache group, so its inventory entry is
retargeted. #9906's new test thread_local is recorded cold — it is a test
file, and the other 23 there are recorded the same way.
