+++ scripts/thread_local_cold_allowlist.json
@@ -1,9 +1,10 @@
 {
   "_comment": "Files still declaring raw `thread_local!`. The count is the number of DECLARATIONS that survive into a shipping build \u2014 each one pays `_tlv_get_addr` per read on Darwin \u2014 and it is a ratchet, so adding a `static` to an already-listed file fails whether or not it opens a new block. New code should use `crate::perry_thread_local!` \u2014 see crates/perry-runtime/src/tls_hot.rs. Regenerate with scripts/check_thread_locals.py --update.",
-  "_hot_declarations": 385,
+  "_hot_declarations": 405,
   "files": {
     "crates/perry-runtime/src/agent.rs": 1,
     "crates/perry-runtime/src/arena/block.rs": 5,
+    "crates/perry-runtime/src/arena/page_meta/mod.rs": 2,
     "crates/perry-runtime/src/async_context.rs": 3,
     "crates/perry-runtime/src/async_hooks.rs": 7,
     "crates/perry-runtime/src/builtins/console.rs": 4,
@@ -54,7 +55,7 @@
     "crates/perry-runtime/src/node_inspector.rs": 2,
     "crates/perry-runtime/src/node_repl.rs": 1,
     "crates/perry-runtime/src/node_stream_constructors.rs": 3,
-    "crates/perry-runtime/src/node_stream_tests.rs": 23,
+    "crates/perry-runtime/src/node_stream_tests.rs": 24,
     "crates/perry-runtime/src/node_submodules/blob.rs": 4,
     "crates/perry-runtime/src/node_submodules/diagnostics.rs": 9,
     "crates/perry-runtime/src/node_submodules/diagnostics_tail.rs": 2,
@@ -83,7 +84,6 @@
     "crates/perry-runtime/src/v8.rs": 2,
     "crates/perry-runtime/src/wasi.rs": 1,
     "crates/perry-runtime/src/weakref.rs": 2,
-    "crates/perry-runtime/src/web_storage.rs": 2,
-    "crates/perry-runtime/src/arena/page_meta/mod.rs": 2
+    "crates/perry-runtime/src/web_storage.rs": 2
   }
 }
